The post methods of cliente and pedidos are ready for the connections between them. Now im gonna do the endpoints for the other clasees

+++ src/main/java/com/example/cafettoapp2/Cliente/ClienteService.java
@@ -48,7 +48,7 @@ public class ClienteService {
 
     // PUT
 
-    public void updateCliente(int clienteId, String name,int phoneNumber,String location) {
+    public void updateCliente(int clienteId, String name,Long phoneNumber,String location) {
 
         Cliente cliente = clienteRepository.findById(clienteId).orElseThrow(() -> new IllegalStateException("No existe el Cliente con id " + clienteId));
